requirements.txt
Flask==1.1.2
gunicorn
dnspython
pandas
pymongo
numpy
python-dotenv
xlwt